構文解析器から独立した3つの責務を分離 (#2364)
nako_parser3.mts (2,934行) のうち、構文規則の相互再帰の外側にあり、
トークンのカーソルにも計算用スタックにも触れない部分を独立モジュールにした。

- core/src/nako_parser_operator.mts (新規)
  中置記法→逆ポーランド→構文木の変換 (infixToPolish / infixToAST)。
  logger を引数に取る自由関数にした。呼び出し元は yGetArgOperator の1箇所のみ。

- core/src/nako_parser_async.mts (新規)
  構文解析後に AST を走査して asyncFn を伝播させる後処理パス (_checkAsyncFn)。
  書き換えの有無を戻り値に畳み込んだので、NakoParserBase.isModifiedNodes を
  削除でき、parse() の不動点ループも1行になった。

- core/src/nako_parser_message.mts (新規)
  エラーメッセージの組み立て (nodeToStr / makeStackBalanceReport)。
  nodeToStr は31箇所から呼ばれるため NakoParserBase に委譲メソッドを残し、
  呼び出し側の差分はゼロにしてある。

- core/test/nako_parser_async_test.mjs (新規)
  分離により単体で検証できるようになった非同期判定のテスト。

行数: nako_parser3.mts 2,934→2,780 / nako_parser_base.mts 414→363

なお Token→Ast 変換 (_tokensToNodes / _tokenToNode) は、yConst と絡む上に
利用箇所が yLet の複数代入構文だけなので、分離せず据え置いた。

検証: tsc エラー0 / lint 警告0 / test:core 856件・test:node 119件・
test:common 30件すべて成功。#2364 で追加した AST ゴールデン比較64件も一致。
